move FeedStoreSpy to a separate helpers file + grouped Feed Cache use cases tests
