+++ EssentialFeedTests/Feed Cache/LoadFeedFromCacheUseCaseTests.swift
@@ -1,0 +1,25 @@
+
+import XCTest
+import EssentialFeed
+
+class LoadFeedFromCacheUseCaseTests: XCTestCase {
+
+    func testInitDoesNotRequestCacheDeletion() {
+        let (_, storeSpy) = makeSUT()
+
+        XCTAssertEqual(storeSpy.messages, [])
+    }
+
+    // MARK: - Helpers
+
+    private func makeSUT(currentDate: @escaping () -> Date = Date.init, file: StaticString = #filePath, line: UInt = #line) -> (LocalFeedLoader, FeedStoreSpy) {
+        let feedStore = FeedStoreSpy()
+        let sut = LocalFeedLoader(store: feedStore, currentDate: currentDate)
+
+        testMemoryLeak(sut, file: file, line: line)
+        testMemoryLeak(feedStore, file: file, line: line)
+
+        return (sut, feedStore)
+    }
+
+}
